Some cleanup from previous commits
