+++ app/javascript/helpers/renderUnsafeXml.js
@@ -1,0 +1,7 @@
+// description: helper to read CSRF token in React components
+
+const renderUnsafeXml = (string) => {
+  return {__html: string};
+}
+
+export default renderUnsafeXml;
